Load astronomical object data from realm

+++ Astro/Abstractions/FormatterMethods.swift
@@ -4,7 +4,7 @@ class FormatterMethods {
     
     let decimalFormatter = NumberFormatter()
     
-    func formatWeight(_ weight: Decimal) -> String {
+    func formatWeight(_ weight: Double) -> String {
         setToTwoDecimalPlaces()
         let stringWeight = decimalFormatter.string(from: weight as NSNumber)
         return "\(stringWeight!) lbs"
